requirements.txt
# 核心框架
streamlit>=1.20.0

# 图像处理与计算机视觉
Pillow>=9.0.0
opencv-python-headless>=4.5.0
numpy>=1.21.0

# 视频合成与音频处理
moviepy>=1.0.3

# 其他工具
tqdm>=4.65.0
